+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.java
@@ -12,8 +12,7 @@ import at.hannibal2.skyhanni.features.bazaar.BazaarBestSellMethod;
 import at.hannibal2.skyhanni.features.bazaar.BazaarOrderHelper;
 import at.hannibal2.skyhanni.features.chat.ChatFilter;
 import at.hannibal2.skyhanni.features.chat.HideFarDeathMessages;
-import at.hannibal2.skyhanni.features.chat.PlayerChatFilter;
-import at.hannibal2.skyhanni.features.chat.SkyBlockLevelChatMessage;
+import at.hannibal2.skyhanni.features.chat.PlayerChatFormatter;
 import at.hannibal2.skyhanni.features.commands.WikiCommand;
 import at.hannibal2.skyhanni.features.damageindicator.DamageIndicatorManager;
 import at.hannibal2.skyhanni.features.dungeon.*;
@@ -80,8 +79,7 @@ public class SkyHanniMod {
 
         registerEvent(new BazaarOrderHelper());
         registerEvent(new ChatFilter());
-        registerEvent(new PlayerChatFilter());
-        registerEvent(new SkyBlockLevelChatMessage());
+        registerEvent(new PlayerChatFormatter());
         registerEvent(new DungeonChatFilter());
         registerEvent(new HideNotClickableItems());
         registerEvent(new ItemDisplayOverlayFeatures());