+++ src/main/java/at/hannibal2/skyhanni/config/Features.java
@@ -8,7 +8,7 @@ import at.hannibal2.skyhanni.config.core.config.Position;
 import at.hannibal2.skyhanni.config.core.config.annotations.Category;
 import at.hannibal2.skyhanni.config.core.config.gui.GuiPositionEditor;
 import at.hannibal2.skyhanni.config.features.*;
-import at.hannibal2.skyhanni.features.chat.PlayerChatFilter;
+import at.hannibal2.skyhanni.features.chat.PlayerChatFormatter;
 import com.google.gson.annotations.Expose;
 import net.minecraft.client.Minecraft;
 
@@ -87,12 +87,12 @@ public class Features extends Config {
         }
 
         if (runnableId.equals("testAllChat")) {
-            PlayerChatFilter.Companion.testAllChat();
+            PlayerChatFormatter.Companion.testAllChat();
             return;
         }
 
         if (runnableId.equals("testGuildChat")) {
-            PlayerChatFilter.Companion.testGuildChat();
+            PlayerChatFormatter.Companion.testGuildChat();
             return;
         }
     }